added watchhelper class and handle edge cases
